Benjam - still not working

+++ First_Step/src/jus/stage/samples/BubbleSort.java
@@ -68,7 +68,7 @@ public class BubbleSort {
 		HashMap<Integer, Long> samples;
 		try {
 			samples = BubbleSort.getSamples(Settings.nbSample);
-			HashMap<long[], Long> featuredSamples = Features.makeFeatures(samples);
+			HashMap<double[], Long> featuredSamples = Features.makeFeatures(samples);
 			return Features.scaling(featuredSamples);
 		} catch (Exception e) {
 			e.getMessage();
